LPS-9180
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@50747 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-service/src/com/liferay/portal/service/persistence/impl/BasePersistenceImpl.java
@@ -68,7 +68,7 @@ public class BasePersistenceImpl<T extends BaseModel<T>>
 			return 0;
 		}
 		else {
-			return ((Integer)results.get(0));
+			return ((Long)results.get(0)).intValue();
 		}
 	}
